// part9/ilari-flight-diary/frontend-flight-diary/src/types.tsx
export enum Weather {
    Sunny = 'sunny',
    Rainy = 'rainy',
    Cloudy = 'cloudy',
    Stormy = 'stormy',
    Windy = 'windy',
}

export enum Visibility {
    Great = 'great',
    Good = 'good',
    Ok = 'ok',
    Poor = 'poor',
}

export interface DiaryEntry {
    id: number;
    comment: string;
    date: string;
    weather: Weather;
    visibility: Visibility;
}

export type NewDiaryEntry = Omit<DiaryEntry, 'id'>;

export type NonSensitiveDiaryEntry = Omit<DiaryEntry, 'comment'>;

export type DiaryEntryToShow = Omit<DiaryEntry, 'id, comment'>;